requirements.txt
streamlit==1.38.0
openai
pyyaml
python-dotenv
pydantic-ai
validators